[REF] split: test_autovacuum
