+++ odoo/addons/test_orm/models/test_autovacuum.py
@@ -1,0 +1,20 @@
+import datetime
+
+from odoo import api, fields, models
+
+
+class TestAutovacuumAutovacuumed(models.Model):
+    _name = 'test_autovacuum.autovacuumed'
+    _description = 'test_autovacuum.autovacuumed'
+
+    expire_at = fields.Datetime('Expires at')
+
+    @api.autovacuum
+    def _gc_simple(self):
+        self.search([('expire_at', '<', datetime.datetime.now() - datetime.timedelta(days=1))]).unlink()
+
+    @api.autovacuum
+    def _gc_proper(self, limit=5):
+        records = self.search([('expire_at', '<', datetime.datetime.now() - datetime.timedelta(days=1))], limit=limit)
+        records.unlink()
+        return len(records), len(records) == limit
