+++ odoo/addons/test_orm/models/test_orm.py
@@ -1,4 +1,3 @@
-import datetime
 import logging
 
 from odoo import api, fields, models
@@ -2387,23 +2386,6 @@ class TestOrmCustomTable_Query_Sql(models.Model):
             GROUP BY tag.id
             """,
         )
-
-
-class TestOrmAutovacuumed(models.Model):
-    _name = 'test_orm.autovacuumed'
-    _description = 'test_orm.autovacuumed'
-
-    expire_at = fields.Datetime('Expires at')
-
-    @api.autovacuum
-    def _gc_simple(self):
-        self.search([('expire_at', '<', datetime.datetime.now() - datetime.timedelta(days=1))]).unlink()
-
-    @api.autovacuum
-    def _gc_proper(self, limit=5):
-        records = self.search([('expire_at', '<', datetime.datetime.now() - datetime.timedelta(days=1))], limit=limit)
-        records.unlink()
-        return len(records), len(records) == limit
 
 
 class TestOrmSharedCompute(models.Model):
