+++ odoo/addons/test_orm/security/test_autovacuum/ir.model.access.csv
@@ -1,0 +1,2 @@
+"id","name","model_id:id","group_id:id","perm_read","perm_write","perm_create","perm_unlink"
+access_test_autovacuum_autovacuumed,access_test_autovacuum_autovacuumed,model_test_autovacuum_autovacuumed,base.group_user,1,0,0,0
