+++ odoo/addons/test_orm/security/test_orm/ir.model.access.csv
@@ -125,7 +125,6 @@ access_test_orm_transient_model,access_test_orm_transient_model,model_test_orm_t
 access_test_orm_custom_view,access_test_orm_custom_view,model_test_orm_custom_view,base.group_user,1,0,0,0
 access_test_orm_custom_table_query,access_test_orm_custom_table_query,model_test_orm_custom_table_query,base.group_user,1,0,0,0
 access_test_orm_custom_table_query_sql,access_test_orm_custom_table_query_sql,model_test_orm_custom_table_query_sql,base.group_user,1,0,0,0
-access_test_orm_autovacuumed,access_test_orm_autovacuumed,model_test_orm_autovacuumed,base.group_user,1,0,0,0
 access_test_orm_shared_compute,access_test_orm_shared_compute,model_test_orm_shared_compute,base.group_user,1,0,0,0
 access_test_orm_model_no_access,access_test_orm_model_no_access,model_test_orm_model_no_access,,0,0,0,0
 access_test_orm_model_all_access_public,access_test_orm_model_all_access_public,model_test_orm_model_all_access,base.group_public,1,0,0,0
